train function and refactor wip
